Move analytics endpoints from history to analytics app
