+++ safe_transaction_service/analytics/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class AnalyticsConfig(AppConfig):
+    name = "safe_transaction_service.analytics"
+    verbose_name = "Analytics for Safe Transaction Service"
