+++ safe_transaction_service/analytics/serializers.py
@@ -1,0 +1,14 @@
+from rest_framework import serializers
+
+from gnosis.eth.django.serializers import EthereumAddressField
+
+
+class AnalyticsMultisigTxsByOriginResponseSerializer(serializers.Serializer):
+    origin = serializers.CharField()
+    transactions = serializers.IntegerField()
+
+
+class AnalyticsMultisigTxsBySafeResponseSerializer(serializers.Serializer):
+    safe = EthereumAddressField()
+    master_copy = EthereumAddressField()
+    transactions = serializers.IntegerField()
