+++ safe_transaction_service/history/tests/test_views.py
@@ -3024,165 +3024,6 @@ class TestViews(SafeTestCaseMixin, APITestCase):
             self.assertEqual(response.status_code, status.HTTP_200_OK)
             self.assertCountEqual(response.data, expected_l2_master_copy)
 
-    def test_analytics_multisig_txs_by_origin_view(self):
-        response = self.client.get(
-            reverse("v1:history:analytics-multisig-txs-by-origin")
-        )
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        origin = "Millennium Falcon Navigation Computer"
-        origin_2 = "HAL 9000"
-        multisig_transaction = MultisigTransactionFactory(origin=origin)
-        response = self.client.get(
-            reverse("v1:history:analytics-multisig-txs-by-origin")
-        )
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        expected = [
-            {"origin": origin, "transactions": 1},
-        ]
-        self.assertEqual(response.data, expected)
-
-        for _ in range(3):
-            MultisigTransactionFactory(origin=origin_2)
-
-        response = self.client.get(
-            reverse("v1:history:analytics-multisig-txs-by-origin")
-        )
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        expected = [
-            {"origin": origin_2, "transactions": 3},
-            {"origin": origin, "transactions": 1},
-        ]
-        self.assertEqual(response.data, expected)
-
-        for _ in range(3):
-            MultisigTransactionFactory(origin=origin)
-
-        # Check sorting by the biggest
-        response = self.client.get(
-            reverse("v1:history:analytics-multisig-txs-by-origin")
-        )
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        expected = [
-            {"origin": origin, "transactions": 4},
-            {"origin": origin_2, "transactions": 3},
-        ]
-        self.assertEqual(response.data, expected)
-
-        # Test filters
-        origin_3 = "Skynet"
-        safe_address = Account.create().address
-        MultisigTransactionFactory(origin=origin_3, safe=safe_address)
-        response = self.client.get(
-            reverse("v1:history:analytics-multisig-txs-by-origin")
-            + f"?safe={safe_address}"
-        )
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        expected = [
-            {"origin": origin_3, "transactions": 1},
-        ]
-        self.assertEqual(response.data, expected)
-
-        response = self.client.get(
-            reverse("v1:history:analytics-multisig-txs-by-origin")
-            + f"?to={multisig_transaction.to}"
-        )
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        expected = [
-            {"origin": multisig_transaction.origin, "transactions": 1},
-        ]
-        self.assertEqual(response.data, expected)
-
-    def test_analytics_multisig_txs_by_safe_view(self):
-        response = self.client.get(reverse("v1:history:analytics-multisig-txs-by-safe"))
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        safe_address_1 = Account.create().address
-        safe_address_2 = Account.create().address
-        safe_address_3 = Account.create().address
-        MultisigTransactionFactory(safe=safe_address_1)
-        MultisigTransactionFactory(safe=safe_address_1)
-        response = self.client.get(reverse("v1:history:analytics-multisig-txs-by-safe"))
-        result = response.json()
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(result["count"], 1)
-        self.assertEqual(
-            result["results"][0],
-            {"safe": safe_address_1, "masterCopy": None, "transactions": 2},
-        )
-        MultisigTransactionFactory(safe=safe_address_1)
-        safe_status_1 = SafeStatusFactory(address=safe_address_1)
-        response = self.client.get(reverse("v1:history:analytics-multisig-txs-by-safe"))
-        result = response.json()
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(result["count"], 1)
-        self.assertIsNotNone(safe_status_1.master_copy)
-        self.assertEqual(
-            result["results"][0],
-            {
-                "safe": safe_address_1,
-                "masterCopy": safe_status_1.master_copy,
-                "transactions": 3,
-            },
-        )
-        MultisigTransactionFactory(safe=safe_address_2)
-        response = self.client.get(reverse("v1:history:analytics-multisig-txs-by-safe"))
-        result = response.json()
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(
-            result["results"],
-            [
-                {
-                    "safe": safe_address_1,
-                    "masterCopy": safe_status_1.master_copy,
-                    "transactions": 3,
-                },
-                {"safe": safe_address_2, "masterCopy": None, "transactions": 1},
-            ],
-        )
-        safe_status_2 = SafeStatusFactory(address=safe_address_2)
-        safe_status_3 = SafeStatusFactory(address=safe_address_3)
-        [MultisigTransactionFactory(safe=safe_address_3) for _ in range(4)]
-        response = self.client.get(reverse("v1:history:analytics-multisig-txs-by-safe"))
-        result = response.json()
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(
-            result["results"],
-            [
-                {
-                    "safe": safe_address_3,
-                    "masterCopy": safe_status_3.master_copy,
-                    "transactions": 4,
-                },
-                {
-                    "safe": safe_address_1,
-                    "masterCopy": safe_status_1.master_copy,
-                    "transactions": 3,
-                },
-                {
-                    "safe": safe_address_2,
-                    "masterCopy": safe_status_2.master_copy,
-                    "transactions": 1,
-                },
-            ],
-        )
-
-        response = self.client.get(
-            reverse("v1:history:analytics-multisig-txs-by-safe")
-            + f"?master_copy={safe_status_1.master_copy}"
-        )
-        result = response.json()
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(
-            result["results"],
-            [
-                {
-                    "safe": safe_address_1,
-                    "masterCopy": safe_status_1.master_copy,
-                    "transactions": 3,
-                },
-            ],
-        )
-
     def test_modules_view(self):
         invalid_address = "0x2A"
         response = self.client.get(
